% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_variables.R
\name{dummies_to_categorical}
\alias{dummies_to_categorical}
\title{Convert Dummy Variables to Categorical Variables}
\usage{
dummies_to_categorical(df, variable_prefix, drop_dummies = TRUE)
}
\arguments{
\item{df}{a data.frame or tibble to modify}

\item{variable_prefix}{a tidy-evaluation style expression which is both the
new variable name to create and the prefix for the existing dummy variables
to convert into the new categorical variable.}

\item{drop_dummies}{(default: true) an indicator to drop the dummy variables
after creating the new categorical variable}
}
\value{
a modified data.frame with a new categorical variable based on the
dummy variables starting with variable_prefix
}
\description{
Convert Dummy Variables to Categorical Variables
}
\seealso{
categories_to_dummies
}
