% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_variables.R
\name{categorical_to_dummies}
\alias{categorical_to_dummies}
\title{Converting Categorical Variables to Dummy Variables and Vice-Versa}
\usage{
categorical_to_dummies(df, variable, drop_categorical = TRUE)
}
\arguments{
\item{df}{the data.frame to modify}

\item{variable}{a tidy-evaluation style expression that names a column to
convert into dummy variables}

\item{drop_categorical}{(default: true) an indicator to remove the original
variable}
}
\value{
a modified data.frame with new columns that are dummy indicators for
each level of the variable specified
}
\description{
This document introduces a pair of functions which should be intuitive to
tidyverse users and standardizes the process for converting categorical
columns of data back and forth from a set of dummy variables (columns of
TRUE/FALSE indicators for each of the categorical levels).
Convert Categorical Variables to Dummy Variables
}
\details{
This function returns a dataframe modified to contain dummy variables
(TRUE/FALSE indicator columns) for each level of the variable specified.

The new column names are automatically constructed to be {variable}_{level}
where variable is given as a function argument and level is specified as
each unique level of the variable given.
}
\seealso{
dummies_to_categorical
}
